Fix opticalwizard for products without storable
The commit 747f9152bf72dfa626cafb426f740123167e3da4 fixed a problem for
products without stock, but broke for products that doesn't control stock.

Make SellableItemSlave.get_remaining_quantity return None when the sellable
doesn't have a storable (that is already considered by it's callsites) and on
the wizard, if the remaining quantity is None, consider all the quantity to be
available.

Change-Id: I569cc2253ebc4587049b20022fb7352ba6e8b58e