Don't reset sellable in SellableItemSteps.
When choosing a sellable to add to a container, do not reset the
sellable search string. Maybe the user wants to change just a part of
it. But when back to the input, selects all the string, so if the user
does not want to reuse it, she will only need to type the next search
string.

Change-Id: I71e02ec71d62729bd906d73439134ead3014ff52
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5547