Only show batches with stock on available
Batches without stock aren't prefilled on the entry, and also it doesn't make
sense to show them all, since the batch numbers will increase from day to day
and we can end up with *lots* of batches on the list that aren't useful.

Create a StorableBatchView.find_available_by_storable to get only results with
stock and use that to populate the list.

Change-Id: I37e63576a4d0e0d2f9aa6b682fc17df1e6b991db