Fix bugs on check payment method slave
Fix a bug were the finish button would still disabled when
MANDATORY_CHECK_NUMBER was set, even if the check number was there.

Workaround: Fix a bug were when the last character of the entry is removed,
kiwi does not update the model anymore, leaving the last valid content as the
value of the model.

Also perform some corrections when adding and deleting bank data.

Change-Id: Ideb853e87484900a12afec996f58995bbea26be3