Avoid confusion when registering initial stock
The initial stock dialog used to get the branch to register from stock app's
branch filter. That's far from obvious (I myself got confused lots of times
thinking it was getting the current branch) and not a good practice.

Instead, show a combo allowing to select the branch to register the initial
stock, and select the current branch by default.

The old .ui file was duplicated so the StockCostDialog can use it without any
modifications.

Change-Id: If49de854d8d486a11a99b127ec6c397655d778a0
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5516