Show branch names the same way on everywhere
There's a difference when accessing branch.person.name and
branch.company.fancy_name because one is the branch's person's name (read as
'Razão Social' on Brazil) and the other is the fancy name (read as 'Nome
Fantasia' on Brazil).

Some places were using the person's name while others, like the PersonField,
the window's title, etc etc where using branch.get_description (that will
return the fancy_name, unless it's empty, in that case it will return the
person's name as a fallback), making it confusing for the user. For instance,
if I was on stock app and wanted to select a filter based on my current branch,
it would be almost impossible for me to know which one to select, since the
fancy name is different from the person's name (for instance, the examples on
stoqlib).

This solves the problem by using get_description on places still using
branch.person.name directly. Also, use it where fancy_name was being used
directly since it provides a fallback in case it's empty/null.

Change-Id: If8692435c34ca84d24f2d6cce0b9e0e5883fe195
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5511