Fix branch description for views
This is the same that we've done for Branch.get_description, but for Views
(since views can't use branch.get_description())

The only exception is BranchView, only used on BranchSearch, because in there
we really want to show the branch's name. But show the fancy name as an
optional column too.

Change-Id: I55e0432c28d3ccc2184b1c415182a5183f557ffc