Fix a test that broke after last kiwi change
Fix a test that broke after commit 604953b1ebaffb9f9daa770412cf9220d6bf5b6a in kiwi.

The code widget is made insensitive when the account type is bank. Since the
test is testing exactly that type, and the widget is insensitive, the changes
to the widget should not be propagated to the model.

The change in kiwi accidentally fixed that bug since, when the account is
created it's code is None. Even with the widget being set to an insensitive
state, it was propagating it's value (u'') to the model, which shouldn't happen

Change-Id: I79a9f4a7b1025112ae1de5f45cb4314790f57e8e