Select the batch directly if it's number is used
This is a usability improvement. On SellableItemSlave subclasses and POS, if a
batch number is scanned, get it's sellable directly.

We can do that because the batch number should be unique among batches, so
having a batch number you have the batch, and having a batch you have the
sellable.

Change-Id: I607af295bfc781daa9e9fcb852561c31a3b0761c