Dont remove objects from the database.
We need to be really carefull when removing objects from the database when
working with synced databases, since that if an object is sent to the
server, and later removed from the client, the server will have inconsistent
data.

This is a paleative change that will clear the reference from an item to
ites parent (a sale item reference to a sale, for instance) to fix the
problem for now. See Bug 5581 for more information

A proper solution would involve registering the object removal and later
propagating that to the other databases.

Change-Id: I32bc3298d5662d5db720013e9f055229ceeec754