Adjust sale discounts on items with quantity of 1
Since we apply the discount percentage on items, they can generate a 3rd
decimal place, that will be rounded to the 2nd, making the value differ.

Find that difference and apply it to a sale item, preferable to one with a
quantity of 1 since, for instance, applying +0,1 to an item with a quantity of
4 would make it's total +0,4 (+0,3 extra than we are trying to adjust here).

We still need to find a way to workaround when there's just one item on the
sale with more than 1 quantity.

Change-Id: I69b2e11c42706d3b7cfccfd14e419bff6248268a