Add a wizard for creating products
Just like personwizard, but for products, offering the option to create a
common product, a product with batches (just a shortcut since it can be changed
later), a product without stock or a consigned one.

This makes it more obvious to the user that those information are a "one time
option" and not something he can change in the future.

This also simplifies the code *a lot* and make the current interface more clean

Change-Id: I84e09b0d8e1de9170e39b8ba1e487dd6f5df5c32