Validate quantities on missing SellableItemSlave users
LoanItemEditor was the only one not considering the case where there could be
duplicated sellables on the same order (in that case, the loan order).

Also, change it's module name to a more obvious one, since there's no
LoanEditor defined there, only the LoanItemEditor.

Improve the TODO message on opticalwizard's _ItemEditor so when it receives
stock validation, the dev known that it should take multiple sellables in
consideration

Change-Id: I34f88ea0544d750814f1058a3ea981d396666b56
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5527