parameters: Add lazy cache
Add a lazy cache that fetches the parameters on demand
when they are used the first time.

Change the api to instantiate the ParameterAccess() as
sysparam when you import the module, that's now possible
with the lazy cache.

Add a __call__() method for backwards compatibility while
we do the migration

Change-Id: Ie1d447a85c2b66130b368c93265c4be6cf8c2286