Check if default method is active
Don't just check if no default method was passed to the constructor, but if
it's on self._widgets too (it would not be if it isn't active or not creatable
for that type of payment).

This fixes an specific issue on the only place using the default method on the
constructor, purchasewizard, when the bill method is not active.

Change-Id: If8115f1a3a921bcec5860cb39bafdc0f81c62d7e