Associates a sellable with a work order.
Now, when creating a work order, you can associate a sellable with that
work order. This means that if the item the client brought to the store
is also an item sold by the store, you can point that.

If the item is not sold by the store, no sellable needs to be associated
with the work order.

Either way, you can also write a description of the item brought by the
client, for details such as color, model, any defects detected, etc.

Change-Id: Id80e8f420948ac58156ede9a163a39e29035f17c
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5575