Use OrderedDict instead of a list of BatchItems
Now that we are already depend on python 2.7 for a lot of things, use
OrderedDict instead of a list of BatchItems.

That makes the object returned by batch selection dialog easier to manipulate
(the mapping is easier to use than to manipulate a list of tuples) and since
they are ordered, they can be used as original batches to keep the order when
editing the batches.

That way we can also use the dict on a domain api. We could not do that with
BatchItem, which was gui specific.

Change-Id: If57901ad5229bf152024df602ad33d7154ca3ace