Fix a small problem when double-clicking on existing batches
If the quantity is not being validated (for example, on POS, where we have a
suggested quantity, but the user can select something above it), the missing
quantity would be 0 (no validation means nothing missing) and that would make
us append 1 everytime a batch was clicked.

Instead of that, allow quantity to be None, so the value there will remain. In
that example, he would be already filled with the suggested quantity passed by
POS to it, it just have to fill the batch number.

Also, add a test that would've caught the problem.

Change-Id: I6c5bced6d5e249bc680ad1c0d41596e46f06d492