Lots of improvements to work order
Add quantity_decreased to WorkOrderITem, just as we already do on SaleItem.
That way we can create the items with the real quantities on optical pre-sales
(the only place that will really use this functionality right know, since
workordereditor won't let add more items than available on stock) and consume
them when we need.

Now, instead of doing the sync_stock magic, it's up to the user to point the
quantity he is consuming (or even returning to stock) in the case of a
pre-sale. On the wo without sale the old behavior is the same (the quantity is
consumed as soon as it's added).

Also, when sending the work order to another branch, don't let that branch
do any stock movimentation.

For now, we are allowing to finish a work order without consuming all items,
just requiring a confirmation. In the future, it should be a parameter.

Change-Id: I66927857a6a4950587cbd10a7644ec2970c07ce1