Fix printing boletos on landscape orientation
boleto is the only report not using weasyprint that sometimes needs to print
in landscape mode (when printing as "carnes").

Add a workaround for it on poppler (since it doesn't allow us to change the
orientation on the gui) and update the sizes to fit both orientations
perfectly on an A4 page.

Change-Id: Ib448afdd9be336d26e5e6af304f31779237d15ab