Add a generate_series database expression
This will allow views to join with GenerateSeries when they need a sequence
of dates to be complete.

For instance, when we create a summary of sales by date, some of the dates
may have no sale, but we still want that date to appear in the results. So
we should use GenerateSeries.

Change-Id: Ib0ea615dad41bdc5bbf0dfbc3bd6aedd03992b43