Fix test_plugins_dialog
A corner case that would fail the test if more than one unofficial plugins was
present on the working tree, and they were next to each other on the ObjectList.

That because when the code removed the first one, the next item is now on the
removed one's place, making the next item be the next's next item.

By duplicating the list we avoid that problem. Also, now we can reenable one
test that should not fail anymore.

Change-Id: I1a4cef891d3369bf5c2d56de080a122f650dbe31