Fix a wrong view test
The test was wrong. It was relying on getting the first payment on the view,
but there are more than one at the point, making it impossible to make sure
that we were getting the right payment.

And at the end, we were not! The test was passing only because of the
behaviour above. When getting the right payment, we can see that the status
is 'Preview' and not 'To Pay' (as it is for all newly created payments)

Change-Id: Iab8c4f21c15241c0256207738c105fa310b33bf3