Organize buttons on SearchDialog better
This makes the buttons better organized. For example, SellableSearch and it's
subclasses now show Ok/Cancel buttons on the right and the extra buttons on the
left, making it look nicer and more alike to other dialogs (e.g. editors)

Also, adjust some paddings so the buttons' spacing looks nicer.

Change-Id: I715b8ab85bc27675e4eb7fe517c8aa43b557544c