Fix a test that was showing a printing dialog
The test broke on some Ubuntu 14.04 machines, a 'C:\\nppdf32Log\\debuglog.txt'
file is being exhaustively created on the stoq directory, and it changed the
print_labels behaviour, making it try to print something and not to show a
warning.

Fix it by not setting the parameter to that file (why it was set to that in
the first place?) and only test print_labels call as a mock (it's not the
wizard's responsibility to test that function, only it's args to it).

Also, improve the test readibility.

Change-Id: I30f79f2ac40256ad1a36987b944db22760a9ba3f