Register a validation function for PersonRoleTypeStep
Without this, by typing an incorrect/incomplete document, it would try to
search for that given person and, since kiwi will return an empty string,
making it match with a lot of persons (any base with more than one person
registered without the document) and breaking ResultSet.one().

By registering a validation function, the user can only click "Forward" if he
has an empty document or a valid one. This also prevents the user typing part
of the string and thinking that it would find it (it will not).

Change-Id: Iec70bc35bc74aac91f9c326de0443934f65bfab2