Avoid tools duplication
This adds a git submodule for utilities. Now we don't need to duplicate pep8,
source-tests, and even the makefile rules.

Also, remove some not used anymore targets and async.mk which doesn't make
sense anymore. The tools directory is still there for the validatecoverage. It
would be nice to put it on stoq-utils too and make it generic.

Change-Id: Ifb617d77eaf28067d548427f5cdc0a72313b514e