Add integration with egg plugins
It is possible to include *.egg plugins. As eggs are a disguised zip
file. This is designed so that no file extraction shall occur.

The *.plugin files are read and parsed by DesktopParser on the fly, as
they are required only to extract some plugin information. Additional
data files inside the egg will be handled by each plugin.

Additionally bootstrap.py file was modified to allow each plugin modify
the way data files inside its' eggs are handled.

A test simulating egg plugin functionality was added in order to
maintain a proper coverage.

Change-Id: Iea6f37878feac2e579af2d80d4a23834930e0086