Move back dependencies on debian/control
The mapping from setup.py to debian/control works, but it doesn't understand
the required version. A 'foo (>=1.0)' would be mapped to just 'python-foo'

Hopefully it will improve in the future!

Also, map twisted to twisted-web. We don't need the whole twisted package
(that depends on a lot of packages), but we can't indicate that on setup.py
(since there's no separation on pypi).

Change-Id: Ia994275548cbf58f8c6827a63b2436a461e3930b