Help differentiating HighjackedViewable
Some tracebacks shows "HighjackedViewable" as the source of some errors. Since
it can be ProductFullStockView or any of it's subclasses, it's difficult to
debug.

This makes every HighjackedViewable be renamed to
"Highjacked<original_class_name>". For example, "ProductFullStockView" when
being used as a HighjackedViewable will have it's name like
"HighjackedProductFullStockView"

Change-Id: If69c55bd0bb616b097cbc3bf3ab97432db91ed8a