Check dependencies before activating a plugin
It is now possible for a plugin to have dependencies.
Dependencies will be specified on *.plugin file as a option
named 'Dependencies'. Multiple dependencies are specified as
names separated by comma.

Dependencies must have prior activation to it's dependent, so
they must be checked on activation and installation - as stoq
initialization will not ensure that independent plugins are
activated first.

Some tests where needed to cover pluginmanager.py properly,
and so they where added. The test simulate an example of a
dependent plugin with a single dependency.

Change-Id: I537b280f9a100e70ba36417470221ead86228b65