Updated stoq Project: stoq-utils  593b8375160c24fd85c5001880d727c5b21ace21
env.sh: Mimic virtualenv behaviour

When sourcing env.sh (from stoq-utils or inside a submodule), we will not
force a new bash to be initiated anymore. Instead, we are doing what virtualenv
doest: We modify PATH/PYTHONPATH/etc and create a function, deactivate, that
can be called to clean up everything.

It's possible to integrate this with virtualenv's activate script. It's just
put a line that will source this file at the end of the script. When calling
deactivate for example, both setups will be reseted.

Change-Id: I4c692c2eea0592c922f194a9b97f67fd06b80c5f