Avoid showing garbage transfer items
When using SYNCHRONIZED_MODE, removing an item from a Transfer, Sale,
WorkOrder, etc, will not remove it from the database, but remove it's order
reference.

Because of that, we can see some "garbage" TransferItems. They were never
transfered so they should never be visible.

Maybe we should make improve the synchronization to remove such items when it's
safe to do so.

Change-Id: I488f9ce82aacc68c687bb9aba4a2667ec1cd30f2
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5669