Workaround a wrong store error
The way it is now, most of the items will come from self.store (so we need to
fetch them to the store we create on checkout) and some from
self._current_store (closed loan items, closed work order items, etc).
Because of that, the comparison can fail.

This is the smallest change we can do to workaround the problem atm, but we
should create self._current_store sooner (ideally when the first
TemporarySaleItem is created to fetch its sellable) and use it. It should also
simplify the code a lot.

Change-Id: I12527652d29f919e8198925b8657cd3045e3de1e