Get cookie data as unicode
The migration of the old cookie if it wasn't in md5 format was not converting
the values of the cookie to unicode, producing an error when when calling
LoginUser.hash.

Instead, always return the cookie data as unicode as it will be used that way.

Change-Id: I60437443432cec3a59d0eca75eb76c5f7592be30
(cherry picked from commit dc75e92075d5ac91deae38ad991f797e8d9ccbba)