Filter daily movement reports on dialog
It is now possible to filter daily movement reports by branch, it comes as a
dialog that display al data within ObjectLists or ObjectTrees. The option
to print the data is still available after the daily movement dialog.

Some optimization was done when querying the database for this report, a
InPaymentView child was created by the name of DailyInPaymentView. The
SalesPerson name, Client name and Sale value queries cost were greatly
mitigated.

Change-Id: Ic3b8b91e041ed3e4a840709c4127c90949f67cc3