Avoid breaking stoq on unique tuple check
We use unique_tuple_exists to check if another tuple exists for that given
values. If there's more than one object there, there's a problem on the
database (that could be solved by UNIQUE constraints), but the tuple exists,
so avoid breaking it.

This could happen for example if more than one sellable exists on the database
with the same code, as we don't have a UNIQUE constraint in that column.

Change-Id: I21650c4c07e28a5e4b21c62d904d1603c7b991d7
(cherry picked from commit 2c7c7afbb357f7cca91a06fb8e79304e7b377213)