Allow purchase for other branch in sync mode
This will let the user create a purchase for another branch when
using stoq in synchronized mode. When doing this, the purchase (and
its payments) will have a negative identifier, and the conector will
give that purchase and payments a definitive one when it arrives at
the destionation branch.

Change-Id: Id419fec8c641c30d6af51cf457a8e6e4874f8d9f
(cherry picked from commit 398696d2f71492efe25787b893c8061a557c55cb)