Add a QuantityColumn for searches
This new column will also display the unit for the sellable.

Also, if the product does not manage stock and the quantity is 0, it will
displayn an infinity symbol instead, to make it more clear to the user that
he can sell it. (So he does not think that the quantity is actually 0)

Change-Id: I3832ecc4c1a2d6402f9ef2225b8f105696688702