Allow to obligate the user to select a salesperson
Change 'ACCEPT_CHANGE_SALESPERSON' to an enum so it can not only
allow/disallow, but obligate the user to choose a salesperson too.

This can be useful when the person who operates the till is not the same as
the salesperson, as in this case the fact that we suggest the current user
would make a lot of sales to be wrongly assigned to that user.

Change-Id: Ifb630c37c074073d501a1cdd8cade5a7004f67e1