Refactor UIForm editor
Allow the user to hide some widgets and set them as not mandatory, using
UIField.

When activating NFe plugin set some widgets back to mandatory and visible,
so we can emmit the NFe correctly.

Also if the NFe plugin is active, dont allow the user to hide or set to not
mandatory some address widgets

Change-Id: Ia389419689f7fe46c53d29246febedb13e16df71