Don't emit ECF for external sales
Accordingly to the ecf documentation, sales done outside of the establishment
doesn't need to emit an ecf coupon:

http://www.pfe.fazenda.sp.gov.br/guia_procedimentos_ecf_06.shtm#8

This adds a new flag to indicate if the sale was done "externally"
(e.g. on an e-commerce website) and considers that on ecf to avoid
creating a coupon when confirming it.

Change-Id: Ic78f555e152f9f82a4901adc74d09855e33f20dd