Make financial refresh more consistent
The refresh method was making a query different from what the searchslave was
doing. More specifically, although they were querying the same view, it was not
considering the date filters, which was making the view ignore it after
editing a transaction (since that triggers a refresh).

Because of that, the refresh done by clicking on the "Search" button and the
refresh() method were not producing the same results on some situations.

Also, it is an unnecessary duplication of the code.

Change-Id: I8e0c400fb76dce3b7c4bdcf42931da9fc291ecfb