Migrate data from patch 05-20
Since 05-20, those domain objects are created with an Invoice object,
and it is expected that they will have it when, for example, we want
to "confirm" them, so migrate the old ones.

Also, avoid a possible error when trying to drop a missing constraint
on some databases.

Change-Id: Id17c2957fb2e0ee2e21638356da1027c74596a07
(cherry picked from commit a55ab86207b6a762e12064ecb07adddc1171fba1)