Filter the sellables to transfer by the selected branch
Because of a join with _StockBranchSummary, the ProductWithStockBranchView
will produce sellables x branch tuples. We are only interested in the one
from the selected branch for the decrease. We can't even let the query return
more than one as .one() will be used.

This is the same solution as the one done in
fa0ac0f565eb613f58ed0af692bdf056cc1079af

Change-Id: I2e6480b5b5b4815d78efcb53fe4e23f1e9fa7adf