Do not export the discount for each item on NF-e txt file.
Instead of export the base price and the discount value on each item. Export
only the final price.
This is necessary because the discount value is not shown on DANFE file
(only the base value). And some people do not accept the NF-e without the real
paid value by them.

Change-Id: Id571c13be0cdb23b592a41d0612d54227a9dd507