Put kiwi on Build-Depends instead of Indep
Indep dependencies are used to build packages for "all" architecture.
Since we changed it to i386/amd64, it would make launchpad not
install kiwi and thus not be able to build Stoq.

And actually, kiwi really belongs on Build-Depends since it is needed
to build for whichever architecture we want to.

Change-Id: I602348bf8e1e66ff573b15b104cc18bf6dc5a25a
(cherry picked from commit e4b4524f47b71219018dce8b9f4c07615d300942)