Modify the base_price only on SellableEditor
If we use price, it will get/set the on_sale_price when the date is
in the on_sale interval. This is confusing because the user will think
that the price the he sets there will be overwritten by the on_sale_price.
Also, he would then try to modify the base price again only to notice that
he would be updating the on_sale_price.

Also, show a label informing the currently price on sale (if it is on sale)
bellow the base price on the editor

Change-Id: I92d3e889ffce1b94fd5e9041ad1154721c36cf5b