Don't filter pending payments out of till daily movement report.
It doesn't make sense to only display paid payments on till daily movements
report. With the exception of money and credit, all other payments will not
me marked as paid automatically, which means that they would not be displayed
on that report, even though they are part of the daily movement.

Change-Id: I5be6d7e409a1ace5c1aa4f5bd8a4cfaf2e547563