Recommend instead of depend on libusb-1.0-0:i386
If multiarch is not enabled, any ':i386' package will not be found and will
make apt-get fail. Also, that package is there to make it easier for people
using SAT on devices that require that 32bit library and is not actually
required for stoq. Leaving it on Recommends is better.

Change-Id: I0e69c32b4202c9ee472a95a82510e59d9fca46b1