Refactor parameter creation process
This changes the way we update the parameters table, removing the necessity
of a database upgrade by the user.

Instead, the parameters will be created on edit. When reading a parameter
that is not defined on the database, the default value will be returned.

Change-Id: Ib2f10c06622cce68130689eb2018685de8d153d5
(cherry picked from commit b921a5fc59bb18da99a1efde5c164263c374b887)