Make it possible to check for person details on insensitive query entry
When the entry is marked insensitive, only set the edit button insensitive
if we don't have an object selected (because it would allow the user to
create and select one). Otherwise, let it sensitive, but open the editor
with visual_mode. This is useful to check for details

Change-Id: I0d51c7bb9dffc8ba247f588861c2c54d8b884902
(cherry picked from commit 97c4b0d3ab1df79a976592e649ecd85b235cbc30)