Workaround missing hostname when creating uri
The dsn cannot contain an empty option for address, which is the case
when the user is connected using the local unix socket.

Make _create_uri workaround the problem itself instead of leaving that
for its callsites, which would mean a code duplication on all of those.

Change-Id: I899338024843284f44ea4fc1f5f541a2e97cc95a
(cherry picked from commit 762c6def98121a66f326fed27afc11da43c7850f)