Fix "enable production"
The init dbadmin command was not pausing stoq-server to create the database.
That is usually ok, since the database doesn't exist and thus stoq-server
should not be running anyway. But when the database exists, it needs to be
dropped first, and if stoq-server is running, it will fail to do that.
That will always happen when the user was using the demo mode and decided
to enable production.

Change-Id: I880c5c3b14213b4d9ed842f7117a8e2aca39fba0