Fix long time bug in payments editor
There is a bug that exists in stoq for a long time related to the attachment
filed in the payment editor.

If you leave the editor open for longer then 5 minutes, when the editor is
confirmed, this exception happens:

  File "/usr/lib/python2.7/dist-packages/stoqlib/gui/editors/baseeditor.py", line 182, in confirm
    slave.on_confirm()
  File "/usr/lib/python2.7/dist-packages/stoqlib/gui/editors/paymenteditor.py", line 170, in on_confirm
    self.model.attachment = self.fields['attachment'].attachment
AttributeError: 'AttachmentField' object has no attribute 'attachment'

Settin attachment as a gobject.property apparently soves the issue.

Change-Id: I903f646d18ab889b333b15be21a599b0470e0289