Set sale client based on document.
This will indicate the sale client if a CPF or CNPJ was informed when opening the ECF
coupon.

This only applies to ECFs that require the document at the coupon opening. Those that
require the document at the coupon closing will already have a client when the document
is added.

Fixes: https://tree.taiga.io/project/stoq/issue/343

Change-Id: Ic3405755ba5c30c795bff107d6d0dfa9a3becd20