Change testsuit to use en_US locale
In python3, the locale C will force the encoding to be ASCII instead
of UTF-8 and that will break all the io operations in the tests.

Python 3.6 is adding the option to coerce that, as they say, legacy locale,
to UTF-8, which means that we could go back to it in the future if we want.
The pep regarding that change is here: https://www.python.org/dev/peps/pep-0538/

Change-Id: Ib16ef7570446f35b353c99800e1d9ae7e608b279