Fix small issue with tax parameter.
We were setting the sellble.tax_constant_id directly from the parameter
value, but if the constant that the parameter references is removed, the
param now references an non existing constant.

This is trying to get the real object that the param references, and if that
does not exist anymore, None is returned, thus not breaking the editor.

Change-Id: I911a03319ad415ecc9b81978db3747bd7453f501