Workaround a new nss issue after py3 migration
When verifying the certificate for the intended usage an exception is thrown warning
that the issuer is not recognized:

nss.error.CertVerifyError: (SEC_ERROR_UNKNOWN_ISSUER) Peer's Certificate issuer is not recognized

This error was already being thrown in py2, but now there is a nasty side effect that
the cert database cannot be shutdown, with the following exception:

File "/home/romaia/stoq/master/stoq/stoqlib/lib/session.py", line 221, in shutdown
  nss.nss_shutdown()
nss.error.NSPRError: (SEC_ERROR_BUSY) NSS could not shutdown. Objects are still in use.

This patch is commenting the verify callback for now and assuming that the
intended usage is valid. (this is the same thing that was happening with python2)

Change-Id: I2b491dbed55dd4b4f1832316e3431ec2aece2799