Workaround gtk3 not respecting the size_request
Since the change to gtk3, it is not respecting the size_request asked by us.
As a workaround, for now, truncate the options available in the tax slaves.
This should not be an issue since the option is prefixed by a code, that
should be enough to identify the option

Change-Id: I9d9eb1b990f83cd894bbfd03e1e62f8d8d3d7b0f