Differentiate package rows in Sale Order Reports
If there are any package products in a sale report the
children products should now appear with a smaller font
size and with some padding.

Also, sales created within the PoS app were created based
on different semantics. Now both PoS and Sales apps behave
the same way - the children indicate the price.

Some reports that had a complete header had to be updated
to not include the subtitle.

Fixes: https://tree.taiga.io/project/stoq/us/795
Change-Id: I622ad03dd400aa214aaf0f0534a9824035ded70a