Fix work order database search
Work order searches raised a programming error because it was
evaluated that they necessarily had a join with the Product table,
which is not true.

The bug happened in a condition statement within the Viewable method
has_join_with. The clause in question actually checked for equality instead
of using the 'is' operator and checking for identity and eliminating the
possibility of false positives.

FIXES: https://tree.taiga.io/project/stoq/us/788
Change-Id: Ie0d96a253d445f2beb64b5bd0842d071e86c86a1