Fix returned sales undo
Generally, returned sales could not be undone because they could violate the
PaymentMethod's max_installments condition. This was fixed by adding a new optional
parameter to the PaymentMethod.create_payment function.

On top of that, the undoing method now tries to cancel a corresponding pending payment related to the
returned sale if it is found. Should that not be the case, a new incoming payment is
made as it was supposed to be before this patch.

Fixes: https://tree.taiga.io/project/stoq/us/827
Change-Id: Ic9aeb7d8a53ae2094d838a1aa2bb128479796f61
(cherry picked from commit 093b3e26459f84931dc4085b24ddbdf895f8651a)