Fix issue transaction rollback to savepoints
When rolling back to a transaction savepoint, there was an issue that objects that were
created inside that savepoint would still be in th store cache - since we implemented
the savepoint infrastructure ourselves.

Some times when flushing the store after the given rollback, storm would try to reload
the object from the database and would end up finding out that the object was removed.

This patch fixes this by explictly listing what objects are created, removed or edited inside
the savepoints and then properly invalidating these objects as storm does on regular
rollbacks.

Fixes: https://tree.taiga.io/project/stoq/us/872
Change-Id: Icbd050f8a1e079a02f05e8aa2266b6f214237f0e