Add parameter to let till be open for a few more hours
For users that don't need an ECF, we should let them keep the till open
for a few more hours after midnight.

Bars and restaurants, for instance, work a few hours past midnight.

Note that if the user is using an ECF, and its past midnight, he will
still be prompted to close the till (since the ECF enforces that).

Fixes: https://tree.taiga.io/project/stoq/task/1074
Change-Id: I35208f53694ff9d5b31d107711f8b30de018c20c