Improve search editor usability
When using a search editor with double_click_confirm=True, when the user
creates a new object and saves it, he would still have to search the
object he just created to select it.

This change is confirming the search (ie, selecting the item) that the
user just created.

Fixes: https://tree.taiga.io/project/stoq/task/1074
Change-Id: I6508a73817450d369e8cd9a3192c146612a47ac8
(cherry picked from commit 2143cb36b0bd9cb254fa8ed15164268737753bbf)