Let default user login automatically.
If there is only one user in the system with username 'admin', and the
password is blank (ie, for a database that was just created), let the
user login without asking a password.

If the user changes the admin password or he creates more users, then
the password prompt will be displayed.

Change-Id: Ic4e8d0727ac2ba1d0dc32b736ccf62f087b0c40a