Fix midnight check on fiscal printer.
There were two issues with the check:

1) It was not destroying the previous check then changing apps. This way, if
the user entered and left the pos app multiple times, a lot of checks would
be created.

2) When recreating the check for the next day, we were calling
setup_midnight_check right after the current one was executed. The problem
is that the current one might run a few miliseconds before midnight, so the
newly added check would run again for the current day.

Fixes: https://tree.taiga.io/project/stoq/task/1258
Change-Id: I4b47058cf63b4355aa37e301602504d0d3431cfd