Add purchase reconciliation wizard
This patch allows to conciliate many receivings that don't have
invoices/payments yet into an unnique invoice, which has it's own payments.
A new table ReceivingInvoice is created to represent
the conciliation, and it can have many receivings related to it. All the
invoice information present on ReceivingOrder was migrated to the new table,
and also the supplier and transporter fields.

Fixes: https://tree.taiga.io/project/stoq/us/1352
Change-Id: I73a9982665c712359af1867a8827d43b7b96416a