Add services app option to create new purchase
Today, a dialog to create a purchase automatically appears when moving
the order from waiting to in progress.

When the work order is already in progress and the user needs a new purchase
order, he can just right click the WO and choose the option to create a
new one.

Fixes: https://tree.taiga.io/project/stoq/task/1578
Change-Id: I607f94de9e05eacdb46077205a19c5adcc14236a