Add event for when a sale is remotelly confirmed.
This event should be emitted when a sale is confirmed by a remote
application (interacting with stoq api via a web service).

This should be used specially by fiscal plugins that should be aware
that the confirmation comes from an outside enviroment, so no gui
operations should be performed.

The event should provide everything that is necessary to emit the fiscal
document (currently the sale that is being confirmed and the client
document).

Fixes: https://tree.taiga.io/project/stoq/task/1599
Change-Id: I710a64b982c7858938d432e85ad3020515e9f56f