Allow payments greater than sale total
With the parameter ALLOW_CREATE_PAYMENT_ON_SALE_QUOTE ative, it is quite common
the total payments be greater than the total quote, since we can edit the
current quote to a lower price (with the payments already registered). When it
happens, the Stoq will only warn the user about this value.

Fixes: https://tree.taiga.io/project/stoq/task/1612

Change-Id: Iaa2f8ee839f26ddce6dd29a35c8b019b30f54284