Try to handle multiple threads better.
autoreload_objects uses a WeakSet to save all currently open stores. But
if a store is closed in another thread while autoreload is going through
all the stores, an exception will be raised that the set changed size
during iteraion.

To aviod the set changing size, create a copy of the set into a list and
iter that list instead.

Fixes: https://trello.com/c/nCJ7IgZ4/95-tratar-erro-de-set-changed-size-during-iteration
Change-Id: I8eb7aec4bb58ed3bd718d22b49055093d7787dc2