Add a new search column in the Till app
The new column added is Missing Payment. This will be very useful for
clients who use the Till app to confirm sales.

In cases where there is partial payment of the sale, it will no longer be
shown only the total value of sales, but also the partial amount to be paid.

Fixes: https://tree.taiga.io/project/stoq/us/2119
Change-Id: I7b05ef60dc16a43fd5ae9c764ed1b2dea9cd901a