Dont break if __repr__ fails.
Sometimes repr(obj) may fail if the object being print is from a closed
connection. This is specially common during crash reports (sentry),
while data is bein collected to send with the report.

This will display an '<database connection closed>" message in the
values that could not be fetched from the database.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/192
Change-Id: I922d82a1736ec10e76f6ae4f0e0b5a1931676e83